Remove Interactive Provider
This is in preparation for "protected mode" for stacks, by allowing the
caller of `update_stack` on the provider to choose whether or not to
force interactive mode.

This also rolls back a small diff when we were preparing to use the
interactive provider when using the `diff` command.
